#3754 Refactor REST configurations
